// src/Storefront.WebApp/app/src/workplace/enums/account/accountType.enum.ts
export enum AccountTypeEnum {
  // Main = 'Main',
  // Additional = 'Additional',

  Undefined = 'Undefined',

  Device1004 = 'Device1004',
  Int2608 = 'Int2608',
  Transit2920 = 'Transit2920',
  Dispute2924 = 'Dispute2924',
  Shortage2924 = 'Shortage2924',
  Surplus2924 = 'Surplus2924',
  Transit2924 = 'Transit2924',
  Fee3570 = 'Fee3570',
  FeeOvd3570 = 'FeeOvd3570',
  Fee3578 = 'Fee3578',
  FeeOvd3578 = 'FeeOvd3578',
  Cashback3678 = 'Cashback3678',
  Limit9129 = 'Limit9129',
  Damping96019618 = 'Damping96019618',
  Reserve9601 = 'Reserve9601',
  Acc2924AdviceOffus = 'Acc2924AdviceOffus',
  Acc3622Ndfl = 'Acc3622Ndfl',
  Acc6050 = 'Acc6050',
  Acc6052 = 'Acc6052',
  Acc2924P2PVb = 'Acc2924P2PVb',
  Acc2924Transit29 = 'Acc2924Transit29',
  Acc3622MilitaryFee = 'Acc3622MilitaryFee',
  Acc3800 = 'Acc3800',
  Acc3801Eur = 'Acc3801Eur',
  Acc3801Usd = 'Acc3801Usd',
  Acc6397 = 'Acc6397',
  Acc6510 = 'Acc6510',
  Acc6510Cash = 'Acc6510Cash',
  Acc6514 = 'Acc6514',
  Acc7020 = 'Acc7020',
  Acc7020Val = 'Acc7020Val',
  Acc7040 = 'Acc7040',
  Acc7040Val = 'Acc7040Val',
  Acc9900 = 'Acc9900',
  Acc9910 = 'Acc9910',
  Over = 'Over',
  OverExpire2809Bpk = 'OverExpire2809Bpk',
  OverInt = 'OverInt',
  OverIntOvd = 'OverIntOvd',
  OverOvd = 'OverOvd',
  OverOvdInt = 'OverOvdInt',
  OverOvdIntOvd = 'OverOvdIntOvd',
  OverRest2809Bpk = 'OverRest2809Bpk',
  Sks = 'Sks',
  SksInt = 'SksInt',
  Асс2924P2P = 'Асс2924P2P',
}